chore: rename inventory components to use kebab-case
