+++ src/client/gui/components/app.tsx
@@ -3,7 +3,7 @@ import { useSelector } from "@rbxts/react-reflex";
 import { selectManaEnabled } from "shared/store/slices/mana/selectors";
 import { DialogueBox } from "./dialogue-box";
 import { ManaBar } from "./fill-bar/mana-bar";
-import { Backpack } from "./inventory/Backpack";
+import { Backpack } from "./inventory/backpack";
 import { Layer } from "./layer";
 import { SilverLogo } from "./silver-logo";
 import { Stats } from "./stats";
