+++ src/client/gui/components/inventory/empty-hotbar-slot.tsx
