+++ src/client/gui/components/inventory/hotbar-frame.tsx
